+++ source/plugins/correlation/correlationplotitem_columnannotations.cpp
@@ -113,7 +113,7 @@ QCPAxis* CorrelationPlotItem::configureColumnAnnotations(QCPAxisRect* axisRect)
     auto* caXAxis = columnAnnotationsAxisRect->axis(QCPAxis::atBottom);
     auto* caYAxis = columnAnnotationsAxisRect->axis(QCPAxis::atLeft);
 
-    auto h = columnAnnotaionsHeight(_columnAnnotationSelectionModeEnabled);
+    auto h = static_cast<int>(columnAnnotaionsHeight(_columnAnnotationSelectionModeEnabled));
     columnAnnotationsAxisRect->setMinimumSize(0, h);
     columnAnnotationsAxisRect->setMaximumSize(QWIDGETSIZE_MAX, h);
